// src/generated/models/LabelTemplate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * LabelTemplate info of a Label template
 */
export type LabelTemplate = {
    color?: string;
    description?: string;
    exclusive?: boolean;
    name?: string;
};